+++ src/main/java/com/openstudies/hibernate/services/common/UserService.java
@@ -18,7 +18,7 @@ public interface UserService {
 
     public User findByEmail(String email);
 
-    public User findById(Long id);
+    public User findById(Integer id);
 
     public void removeById(Long id);
